+++ kafka-admin/src/main/java/io/strimzi/admin/kafka/admin/AdminClientWrapper.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.strimzi.admin.common.data.fetchers;
+package io.strimzi.admin.kafka.admin;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
