+++ kafka-admin/src/main/java/io/strimzi/admin/kafka/admin/handlers/TopicCreateHandler.java
@@ -4,13 +4,11 @@
  */
 package io.strimzi.admin.kafka.admin.handlers;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
 import io.netty.handler.codec.http.HttpResponseStatus;
-import io.strimzi.admin.common.data.fetchers.AdminClientWrapper;
-import io.strimzi.admin.common.data.fetchers.TopicOperations;
-import io.strimzi.admin.common.data.fetchers.model.Types;
+import io.strimzi.admin.kafka.admin.AdminClientWrapper;
+import io.strimzi.admin.kafka.admin.TopicOperations;
+import io.strimzi.admin.kafka.admin.model.Types;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
@@ -60,9 +58,7 @@ public class TopicCreateHandler extends CommonHandler {
         return routingContext -> {
             setOAuthToken(acConfig, routingContext);
             Future<AdminClientWrapper> acw = createAdminClient(vertx, acConfig);
-
             Types.NewTopic inputTopic = new Types.NewTopic();
-
             Promise<Types.NewTopic> prom = Promise.promise();
             ObjectMapper mapper = new ObjectMapper();
             try {
@@ -74,24 +70,7 @@ public class TopicCreateHandler extends CommonHandler {
             }
 
             TopicOperations.createTopic(acw, prom, inputTopic);
-
-            prom.future().onComplete(res -> {
-                if (res.failed()) {
-                    routingContext.response().setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code());
-                    routingContext.response().end(res.cause().getMessage());
-                } else {
-                    ObjectWriter ow = new ObjectMapper().writer().withDefaultPrettyPrinter();
-                    String json = null;
-                    try {
-                        json = ow.writeValueAsString(res.result());
-                    } catch (JsonProcessingException e) {
-                        routingContext.response().setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code());
-                        routingContext.response().end(e.getMessage());
-                    }
-                    routingContext.response().setStatusCode(HttpResponseStatus.OK.code());
-                    routingContext.response().end(json);
-                }
-            });
+            processResponse(prom, routingContext);
         };
     }
 }
