+++ kafka-admin/src/main/java/io/strimzi/admin/kafka/admin/handlers/TopicUpdateHandler.java
@@ -4,13 +4,11 @@
  */
 package io.strimzi.admin.kafka.admin.handlers;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
 import io.netty.handler.codec.http.HttpResponseStatus;
-import io.strimzi.admin.common.data.fetchers.AdminClientWrapper;
-import io.strimzi.admin.common.data.fetchers.TopicOperations;
-import io.strimzi.admin.common.data.fetchers.model.Types;
+import io.strimzi.admin.kafka.admin.AdminClientWrapper;
+import io.strimzi.admin.kafka.admin.TopicOperations;
+import io.strimzi.admin.kafka.admin.model.Types;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
@@ -58,9 +56,7 @@ public class TopicUpdateHandler extends CommonHandler {
         return routingContext -> {
             setOAuthToken(acConfig, routingContext);
             Future<AdminClientWrapper> acw = createAdminClient(vertx, acConfig);
-
             Types.UpdatedTopic updatedTopic = new Types.UpdatedTopic();
-
             Promise<Types.UpdatedTopic> prom = Promise.promise();
             ObjectMapper mapper = new ObjectMapper();
             try {
@@ -72,23 +68,7 @@ public class TopicUpdateHandler extends CommonHandler {
             }
 
             TopicOperations.updateTopic(acw, updatedTopic, prom);
-            prom.future().onComplete(res -> {
-                if (res.failed()) {
-                    routingContext.response().setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code());
-                    routingContext.response().end(res.cause().getMessage());
-                } else {
-                    ObjectWriter ow = new ObjectMapper().writer().withDefaultPrettyPrinter();
-                    String json = null;
-                    try {
-                        json = ow.writeValueAsString(res.result());
-                    } catch (JsonProcessingException e) {
-                        routingContext.response().setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code());
-                        routingContext.response().end(e.getMessage());
-                    }
-                    routingContext.response().setStatusCode(HttpResponseStatus.OK.code());
-                    routingContext.response().end(json);
-                }
-            });
+            processResponse(prom, routingContext);
         };
     }
 }
